backend-engineer/requirements.txt
fastapi
uvicorn
asyncpg
aioredis
pika
python-jose[cryptography]
python-dotenv
passlib[bcrypt]
sqlalchemy[asyncio]
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
prometheus-fastapi-instrumentator
python-json-logger
bcrypt==4.0.1
pytest
pytest-asyncio
httpx
coverage